Fix for Client key exchange algoritm selection
A client now supports all key exchange algorithms (hence indicating
support in the supported_groups extension). The selected algorithms in
the configuration of the endpoint is now placed at the first spot in the
list, resulting in this algorithm being the key_share of the
ClientHello. If the server doesn't support this group, the proper TLS
retry mechanism is executed to select another algorithm from the
supported list.